64-bit windows was missing
hg-changeset: dea420e2bb2e9216ab4555b8ce84120c1910e916

+++ cimple/wmi/WMI_Adapter.cpp
@@ -1509,7 +1509,7 @@ int cimple_wmi_adapter(
 
     CIMPLE_ENTER;
 
-    long function = (long)arg1;
+    long function = (long)(intptr_t)arg1;
 
     // Dispatch operation:
